Lightbox: Multi-Image Support
Now, each item in the portfolio can display one *or more* images when
clicked. If the data source contains 2 or more images for a given
gallery item, a new row is output into the modal containing a responsive
strip of thumbnails, which load their full-size counterpart into the
modal image area on click. This feature was preceded by some other major
changes, including a switch to a JS-based solution for parsing the JSON
blob of portfolio data and displaying it in the lightbox modal.